Update: 角度に関する処理を別クラスに分ける
