+++ GpsMap/GpsMap_2/ViewController.swift
@@ -36,13 +36,15 @@ class ViewController: UIViewController, CLLocationManagerDelegate, UITextFieldDe
     let speech = AVSpeechSynthesizer()
     var stepCount = 0
     var prevCoordinateInfo: CLLocation? = nil
-    let setAngle: Float = 15.0
+    
     // 現在地ボタン
     var button2 = UIButton()
     let image = UIImage(named: "arrow")
     // 音声テキストボタン
     var button3 = UIButton()
     let image2 = UIImage(named: "mic")
+    
+    let directionJudge = DirectionJudge()
    
 //    @IBOutlet var button: [UIButton]!
     // マイクの変数
@@ -190,11 +192,11 @@ class ViewController: UIViewController, CLLocationManagerDelegate, UITextFieldDe
         
         let nextLocation = self.step.steps[self.stepCount]
         // 目標角度
-        let targetRadian = self.angle(coordinate: prevCoordinateInfo!.coordinate, coordinate2: nextLocation.polyline.coordinate)
-        let targetRadian2 = self.angle(coordinate: nextLocation.polyline.coordinate, coordinate2: prevCoordinateInfo!.coordinate)
+        let targetRadian = directionJudge.angle(coordinate: prevCoordinateInfo!.coordinate, coordinate2: nextLocation.polyline.coordinate)
+        let targetRadian2 = directionJudge.angle(coordinate: nextLocation.polyline.coordinate, coordinate2: prevCoordinateInfo!.coordinate)
         // 実際に移動した角度
-        let userRadian = self.angle(coordinate: prevCoordinateInfo!.coordinate, coordinate2: locations.last!.coordinate )
-        let userRadian2 = self.angle(coordinate: locations.last!.coordinate, coordinate2: prevCoordinateInfo!.coordinate)
+        let userRadian = directionJudge.angle(coordinate: prevCoordinateInfo!.coordinate, coordinate2: locations.last!.coordinate )
+        let userRadian2 = directionJudge.angle(coordinate: locations.last!.coordinate, coordinate2: prevCoordinateInfo!.coordinate)
         
         print("前回の位置座標\(prevCoordinateInfo!.coordinate)")
         print("現在の位置座標\(locations.last!.coordinate)")
@@ -208,28 +210,28 @@ class ViewController: UIViewController, CLLocationManagerDelegate, UITextFieldDe
         }
         
         // 角度の評価を行う
-        if targetRadian - setAngle < 0 || targetRadian + setAngle > 360 {
+        if targetRadian - directionJudge.setAngle < 0 || targetRadian + directionJudge.setAngle > 360 {
             
-           Flag = Flag || self.compareAngle(targetRadian: targetRadian, userRadian: userRadian)
-           Flag = Flag || self.compareAngle(targetRadian: userRadian, userRadian: targetRadian)
+           Flag = Flag || directionJudge.compareAngle(targetRadian: targetRadian, userRadian: userRadian)
+           Flag = Flag || directionJudge.compareAngle(targetRadian: userRadian, userRadian: targetRadian)
         } else {
             
-           Flag = Flag || self.compareAngle2(targetRadian: targetRadian, userRadian: userRadian)
-           Flag = Flag || self.compareAngle(targetRadian: userRadian, userRadian: targetRadian)
+           Flag = Flag || directionJudge.compareAngle2(targetRadian: targetRadian, userRadian: userRadian)
+           Flag = Flag || directionJudge.compareAngle(targetRadian: userRadian, userRadian: targetRadian)
         }
         
-        if targetRadian2 - setAngle < 0 || targetRadian2 + setAngle > 360 {
+        if targetRadian2 - directionJudge.setAngle < 0 || targetRadian2 + directionJudge.setAngle > 360 {
             
-           Flag = Flag || self.compareAngle(targetRadian: targetRadian2, userRadian: userRadian2)
-            Flag = Flag || self.compareAngle(targetRadian: userRadian2, userRadian: targetRadian2)
+            Flag = Flag || directionJudge.compareAngle(targetRadian: targetRadian2, userRadian: userRadian2)
+            Flag = Flag || directionJudge.compareAngle(targetRadian: userRadian2, userRadian: targetRadian2)
         } else {
             
-           Flag = Flag || self.compareAngle2(targetRadian: targetRadian2, userRadian: userRadian2)
-           Flag = Flag || self.compareAngle2(targetRadian: userRadian2, userRadian: targetRadian2)
+           Flag = Flag || directionJudge.compareAngle2(targetRadian: targetRadian2, userRadian: userRadian2)
+           Flag = Flag || directionJudge.compareAngle2(targetRadian: userRadian2, userRadian: targetRadian2)
         }
         
         // 距離の評価を行う
-        let currentDistance = self.distance(current: ( currentCoordinate.latitude, currentCoordinate.longitude), target: (nextLocation.polyline.coordinate.latitude, nextLocation.polyline.coordinate.longitude))
+        let currentDistance = directionJudge.distance(current: ( currentCoordinate.latitude, currentCoordinate.longitude), target: (nextLocation.polyline.coordinate.latitude, nextLocation.polyline.coordinate.longitude))
         let targetDistance = nextLocation.distance
         
         if currentDistance  > targetDistance + 5.0 {
@@ -255,121 +257,15 @@ class ViewController: UIViewController, CLLocationManagerDelegate, UITextFieldDe
     // 磁気センサからユーザーの角度を取得
     func locationManager(_ manager: CLLocationManager, didUpdateHeading newHeading: CLHeading) {
         // ユーザの向いている方向
-        _ = self.degToRad(degrees: (self.mapView.camera.heading))
+        _ = directionJudge.degToRad(degrees: (self.mapView.camera.heading))
         print("カメラ角度")
         print(mapView.camera.heading)
         print("-------------------------------------")
     }
     
-    // 座標から距離を求める関数（三角球面法）
-    func distance(current: (la: Double, lo: Double), target: (la: Double, lo: Double)) -> Double {
-        
-        // 緯度経度をラジアンに変換
-        let currentLa   = current.la * Double.pi / 180
-        let currentLo   = current.lo * Double.pi / 180
-        let targetLa    = target.la * Double.pi / 180
-        let targetLo    = target.lo * Double.pi / 180
-
-        // 赤道半径
-        let equatorRadius = 6378137.0
-        
-        // 算出
-        let averageLat = (currentLa - targetLa) / 2
-        let averageLon = (currentLo - targetLo) / 2
-        let distance = equatorRadius * 2 * asin(sqrt(pow(sin(averageLat), 2) + cos(currentLa) * cos(targetLa) * pow(sin(averageLon), 2)))
-        return distance
-    }
-    
     // 角度に関する関数
     func rotateManager(heading: CLLocationDirection) {
         self.mapView.camera.heading = heading
-    }
-    
-    // 角度をラジアンに変換する
-    func degToRad(degrees: CGFloat) -> CGFloat {
-        return degrees * CGFloat.pi / 180
-    }
-    
-    // 各位を計算
-    func angle(coordinate: CLLocationCoordinate2D, coordinate2: CLLocationCoordinate2D) -> Float {
-        let currentLatitude     = degToRad(degrees: coordinate.latitude)
-        let currentLongitude    = degToRad(degrees: coordinate.longitude)
-        let targetLatitude      = degToRad(degrees: coordinate2.latitude)
-        let targetLongitude     = degToRad(degrees: coordinate2.longitude)
-        
-        let difLongitude = targetLongitude - currentLongitude
-        let y = sin(difLongitude)
-        let x = cos(currentLatitude) * tan(targetLatitude) - sin(currentLatitude) * cos(difLongitude)
-        let p = atan2(y, x) * 180 / CGFloat.pi
-        
-        if p < 0 {
-            return Float(360 + atan2(y, x) * 180 / CGFloat.pi)
-        }
-        return Float(atan2(y, x) * 180 / CGFloat.pi)
-    }
-    
-    // 角度を比較し、アナウンスするか否かの処理(０と３６０の間をまたぐとき）
-    func compareAngle(targetRadian: Float, userRadian: Float) -> Bool{
-        // １つ目の計算用変数の角度調整
-        var calculationRadian = targetRadian + setAngle
-        
-        if calculationRadian > 360  {
-            
-            calculationRadian -= 360
-            
-        }
-        
-        // ２つ目の計算用変数の角度調整
-        var calculationRadian2 = targetRadian - setAngle
-        
-        if calculationRadian2 < 0 {
-            
-            calculationRadian2 += 360
-            
-        }
-            
-        if userRadian < calculationRadian || userRadian > calculationRadian2 {
-            
-            let a  = true
-            return a
-            
-        } else {
-            let a  = false
-            return a
-        }
-    }
-    
-    // 角度を比較し、アナウンスするか否かの処理(０と３６０の間をまたがいないとき)
-    func compareAngle2(targetRadian: Float, userRadian: Float) -> Bool {
-     
-        // １つ目の計算用変数の角度調整
-        var calculationRadian = targetRadian + setAngle
-        
-        if calculationRadian > 360 {
-            
-            calculationRadian -= 360
-            
-        }
-        // ２つ目の計算用変数の角度調整
-        var calculationRadian2 = targetRadian - setAngle
-        
-        if calculationRadian2 < 0 {
-            
-            calculationRadian2 += 360
-            
-        }
-            
-        if userRadian < calculationRadian && userRadian > calculationRadian2 {
-            
-            let a = true
-            return a
-            
-        } else {
-            
-            let a = false
-            return a
-            
-        }
     }
     
     // 画面の初期位置の設定
